refactor(memory): procedural-memory-extractor.ts 분해 — composition 패턴 (#634)
719줄 단일 파일을 책임별 3개 모듈로 분해:
- procedural-memory-field-extractors.ts (429줄): 필드 추출 함수
- procedural-memory-similarity.ts (266줄): 유사도·병합 전략
- procedural-memory-extractor.ts (54줄): 오케스트레이터 (re-export + RuleBasedProceduralExtractor)

기존 import 경로·export surface 변경 없음. 전체 테스트 통과.

Closes #634
